data_collection/pyproject.toml
[tool.poetry]
name = "ai-agents-env-setup-data-collection"
version = "0.1.0"
description = "Code associated with data collection from Environment Setup project."
authors = ["Alexandra Eliseeva <alexandra.eliseeva@jetbrains.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
tqdm = "^4.66.4"
python-dotenv = "^1.0.1"
hydra-core = "^1.3.2"
pandas = "^2.2.2"
gitpython = "^3.1.43"
pydantic = "^2.7.4"
jsonlines = "^4.0.0"

[tool.poetry.group.dev.dependencies]
mypy = "^1.5.0"
pytest = "^7.4.0"
ruff = "^0.3.2"
pyright = "^1.1.367"
isort = "^5.13.2"
pre-commit = "^3.7.1"

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
extend-select = ["I"]

[tool.isort]
profile = "black"
force_sort_within_sections = true
order_by_type = true

[tool.mypy]
python_version = "3.9"

[[tool.mypy.overrides]]
module = []
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"